pull in latest puppet-lint plugins
To ensure that we pull in the latest puppet-lint plugins, we can require their latest version. Background: It's a bit hacky in pdk to update existing gems. The suggest way is to raise the gem version in the Gemfile. Now adding a dozen gems with their version to a Gemfile would make voxpupuli-puppet-lint-plugins obsolete. I think it makes sense that we from time to time update the version requirements in voxpupuli-puppet-lint-plugins.