Add puppet-lint-exec_idempotency-check
What does the plugin do?

When using `exec` resource it is highly recommended to take care on idempotency.
That means, that we need a check to verify if the command should be run again or is required to run at all.

The following attributes control this behavior:

- `creates`: checks if a directory or file exists
- `onlyif` or `unless`: execute checks to verify idempotency
- `refreshonly`: only run the exec if it is triggered by another reosurce

This is part of the puppet best practices since a long time, we just
never had a linter to verify it.