Add puppet-lint-params_not_optional_with_undef-check
This adds a puppet-lint plugin to check for parameters without Optional as type but `undef` as value.

According the the Vox Pupuli best practices, a class parameter with the string datatype should default to `undef` and not `''`, if it's `Optional[]`.
The recommendations are documented at [voxpupuli.org](https://voxpupuli.org/docs/reviewing_pr/).